requirements.txt
requests==2.26.*
python-dotenv==0.19.*
